moved backendClient from Database to new directory under Loterry
